Issue's dialogs helps wizard CBP
